requirements.txt
wagtail==2.11.3
psycopg2==2.8.6
psycopg2-binary==2.8.5
django-countries==7.0
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-pwa==1.0.10
django-recaptcha==2.0.6
dj-rest-auth==2.1.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
django-extensions==3.1.0
django-comments-xtd==2.8.1
wagtailfontawesome==1.2.1
python-decouple==3.4
PyJWT==1.7.1
django-import-export==2.5.0
wagtail-localize==0.9.5
gunicorn
whitenoise
dj-static
dj-database-url
django-toolbelt